requirements.txt
fastapi
uvicorn
pydantic
PyPDF2
python-docx
requests
beautifulsoup4
wikipedia-api
faiss-cpu
numpy
sentence-transformers
transformers
torch
